# docs/OPENAPI_PRODUCTION.yaml
openapi: 3.1.0
info:
  title: PROVENANCE VERIFIED™ Production Authority API
  version: 8.1.3-r3
  description: >-
    Fail-closed tenant-scoped trust authority. Sandbox is deterministic and isolated.
    Pilot is real but non-authoritative. Production issuance requires a valid signed activation record.
servers:
  - url: https://api.provenanceverified.org
    description: Production authority gateway
  - url: https://pilot-api.provenanceverified.org
    description: Pilot non-authoritative authority gateway
security:
  - bearerAuth: []
    tenantHeader: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT-or-scoped-client-secret }
    tenantHeader: { type: apiKey, in: header, name: x-provenance-tenant }
  parameters:
    IdempotencyKey: { in: header, name: idempotency-key, required: true, schema: { type: string, minLength: 16 } }
...
paths:
...
  /api/v1/status:
    get:
      security: []
      summary: Dependency-aware public readiness
      responses: { '200': { description: Ready }, '503': { description: Critical dependency unavailable } }
...
  /api/v1/webhooks/replay:
    post: { parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], responses: { '200': { description: Attributed replay queued }, '409': { description: Replay denied } } }
...